+++ FrameBlend/FrameBlend.cpp
@@ -366,7 +366,7 @@ private:
 
             for (int x = procWindow.x1; x < procWindow.x2; x++) {
                 const PIX *srcPix = (const PIX *)  (_srcImg ? _srcImg->getPixelAddress(x, y) : 0);
-                size_t renderPix = ( (_renderWindow.x2 - _renderWindow.x1) * (y - _renderWindow.y1) +
+                size_t renderPix = ( (size_t)(_renderWindow.x2 - _renderWindow.x1) * (y - _renderWindow.y1) +
                                      (x - _renderWindow.x1) );
                 int count = _countData ? _countData[renderPix] : 0;
                 float sumWeights = _sumWeightsData ? _sumWeightsData[renderPix] : 0;
@@ -753,7 +753,7 @@ FrameBlendPlugin::setupAndProcess(FrameBlendProcessorBase &processor,
     }
 
     const OfxRectI& renderWindow = args.renderWindow;
-    size_t nPixels = (renderWindow.y2 - renderWindow.y1) * (renderWindow.x2 - renderWindow.x1);
+    size_t nPixels = (size_t)(renderWindow.y2 - renderWindow.y1) * (renderWindow.x2 - renderWindow.x1);
     OperationEnum operation = processor.getOperation();
 
     // Main processing loop.
